Crate rustic_meta

Structs

FilterInfo

Static metadata describing a filter type.

FilterInput

One connectable input port off a filter Port index = position in FilterInfo::inputs

MetaGenerator

MetaSink

Enums

ListSize

Determines the size of a List parameter.

Literal

Represents a literal value that defines a List parameter’s type.

MixMode

Strategy for combining multiple audio blocks arriving at the same input port.

Parameter

Traits

MetaFilter

Trait for filters that support named parameter modification and metadata. Implemented automatically by the FilterMetaData derive macro.